requirements.txt
discord.py==1.3.2
psycopg2==2.8.4
requests==2.23.0